vibe: ツールバーをモバイル幅でハンバーガーメニュー化
useIsMobile()でデスクトップ/モバイルを分岐。モバイル幅では
Lucky/Palette/POIトグルを独立表示し、Share/Save/Supersampling/
Jump/SettingsはDropdownMenuに集約。Debugトグルは非表示。

ダイアログ状態 (share/ss/jump/settings) はToolbarに集約し、
ShareDialogHostとSupersamplingDialogを新設。SupersamplingForm
を抽出してPopover(デスクトップ)とDialog(モバイル)で共有。
